Completed Modification Table
All that is left is ensuring that all other portions of the code in which the guns are used use the scriptable objects to ensure that the guns stats change whenever the player modifies their weapon. Additionally there will need to be a call to reset the gun's stats at the beginning and end of the game to ensure that the stats return to their original values.